+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/GeneralSettingsController.java
@@ -29,6 +29,8 @@ import java.util.stream.Collectors;
 import com.tesshu.jpsonic.command.GeneralSettingsCommand;
 import com.tesshu.jpsonic.domain.system.IndexScheme;
 import com.tesshu.jpsonic.feature.i18n.ServerLocaleService;
+import com.tesshu.jpsonic.feature.theme.ServerThemeService;
+import com.tesshu.jpsonic.feature.theme.Theme;
 import com.tesshu.jpsonic.infrastructure.settings.SKeys;
 import com.tesshu.jpsonic.infrastructure.settings.SettingsFacade;
 import com.tesshu.jpsonic.persistence.core.entity.User;
@@ -37,8 +39,6 @@ import com.tesshu.jpsonic.service.MusicIndexService;
 import com.tesshu.jpsonic.service.ScannerStateService;
 import com.tesshu.jpsonic.service.SecurityService;
 import com.tesshu.jpsonic.service.ShareService;
-import com.tesshu.jpsonic.theme.ServerThemeService;
-import com.tesshu.jpsonic.theme.Theme;
 import com.tesshu.jpsonic.util.PathValidator;
 import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.stereotype.Controller;
