+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/PersonalSettingsController.java
@@ -36,6 +36,8 @@ import com.tesshu.jpsonic.domain.system.AvatarScheme;
 import com.tesshu.jpsonic.domain.system.SpeechToTextLangScheme;
 import com.tesshu.jpsonic.domain.system.SupportableBCP47;
 import com.tesshu.jpsonic.feature.i18n.ServerLocaleService;
+import com.tesshu.jpsonic.feature.theme.ServerThemeService;
+import com.tesshu.jpsonic.feature.theme.Theme;
 import com.tesshu.jpsonic.infrastructure.settings.SKeys;
 import com.tesshu.jpsonic.infrastructure.settings.SettingsFacade;
 import com.tesshu.jpsonic.persistence.core.entity.User;
@@ -43,8 +45,6 @@ import com.tesshu.jpsonic.persistence.core.entity.UserSettings;
 import com.tesshu.jpsonic.service.AvatarService;
 import com.tesshu.jpsonic.service.SecurityService;
 import com.tesshu.jpsonic.service.ShareService;
-import com.tesshu.jpsonic.theme.ServerThemeService;
-import com.tesshu.jpsonic.theme.Theme;
 import jakarta.servlet.http.HttpServletRequest;
 import org.apache.commons.lang3.StringUtils;
 import org.springframework.stereotype.Controller;
