+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/controller/StatusChartController.java
@@ -30,9 +30,9 @@ import java.util.Collections;
 import java.util.Date;
 import java.util.List;
 
+import com.tesshu.jpsonic.feature.theme.ChartColor;
 import com.tesshu.jpsonic.service.StatusService;
 import com.tesshu.jpsonic.service.StatusService.TransferStatus;
-import com.tesshu.jpsonic.theme.ChartColor;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.jfree.chart.ChartFactory;
