+++ jpsonic-main/src/main/java/com/tesshu/jpsonic/feature/theme/ChartColor.java
@@ -14,31 +14,28 @@
  * You should have received a copy of the GNU General Public License
  * along with this program. If not, see <http://www.gnu.org/licenses/>.
  *
- * (C) 2009 Sindre Mehus
- * (C) 2016 Airsonic Authors
- * (C) 2018 tesshucom
+ * (C) 2025 tesshucom
  */
 
-package com.tesshu.jpsonic.theme;
+package com.tesshu.jpsonic.feature.theme;
+
+import java.awt.Color;
 
 /**
- * Contains the ID and name for a theme.
+ * Default chart colors used across all chart types.
  */
-public class Theme {
+public enum ChartColor {
+    BACKGROUND(new Color(0xFF, 0xFF, 0xFF)), // ffffff
+    TEXT(new Color(0x33, 0x33, 0x33)), // 333333
+    STROKE(new Color(0x79, 0xA2, 0xD4)); // 79a2d4
 
-    private final String id;
-    private final String name;
+    private final Color color;
 
-    public Theme(String id, String name) {
-        this.id = id;
-        this.name = name;
+    ChartColor(Color color) {
+        this.color = color;
     }
 
-    public String getId() {
-        return id;
-    }
-
-    public String getName() {
-        return name;
+    public Color get() {
+        return color;
     }
 }
