+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/feature/theme/CssUrlProviderImplTest.java
@@ -17,7 +17,7 @@
  * (C) 2026 tesshucom
  */
 
-package com.tesshu.jpsonic.theme;
+package com.tesshu.jpsonic.feature.theme;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.mockito.Mockito.when;
