+++ jpsonic-main/src/test/java/com/tesshu/jpsonic/infrastructure/settings/SettingsFacadeBuilder.java
@@ -36,10 +36,10 @@ import java.util.Collections;
 import java.util.List;
 
 import com.tesshu.jpsonic.feature.i18n.I18nSKeys;
+import com.tesshu.jpsonic.feature.theme.ThemeSKeys;
 import com.tesshu.jpsonic.infrastructure.core.EnvironmentProvider;
 import com.tesshu.jpsonic.persistence.DBSKeys;
 import com.tesshu.jpsonic.service.upnp.UPnPSKeys;
-import com.tesshu.jpsonic.theme.ThemeSKeys;
 import org.apache.commons.configuration2.Configuration;
 import org.apache.commons.configuration2.FileBasedConfiguration;
 import org.apache.commons.configuration2.PropertiesConfiguration;
